feat: rename dateFormat.js to format.js and add formatPhoneNumber function
