+++ src/components/Activity/ActivityInfo.jsx
@@ -1,5 +1,5 @@
 import styles from "./Activity.module.scss";
-import { formatDate, formatTime } from "@/utils/dateFormat";
+import { formatDate, formatTime } from "@/utils/format";
 import { useActivity } from "@/contexts/ActivityContext";
 
 const ActivityInfo = () => {
